refactor(native): centralize CEF bridge client logic
