+++ proton/internal/native/ffi/moon.pkg
@@ -20,10 +20,12 @@ options(
     "src/proton_system_path.c",
     "src/proton_update.c",
     "src/proton.c",
+    "src/engine/cef_common/bridge_client.c",
     "src/engine/cef_common/bridge_lifecycle.c",
     "src/engine/cef_common/browser_lifecycle.c",
     "src/engine/cef_common/browser_session.c",
     "src/engine/cef_common/bridge_policy.c",
+    "src/engine/cef_common/bridge_request.c",
     "src/engine/cef_common/bridge_renderer.c",
     "src/engine/cef_common/cookie_cache.c",
     "src/engine/cef_common/image.c",
